selftest tui: fold secret.go into prompt.go
cmdio.Secret is a tiny wrapper over the same prompt machinery; keeping a
dedicated file for one 13-line function added more navigation overhead
than it saved.
